+++ src/context.ts
@@ -29,6 +29,7 @@ export const GANTT_DEFAULTS = {
   draggable: false,
   rowMovable: false,
   resizable: false,
+  progressDraggable: false,
   linkable: false,
   snapToGrid: false,
   dragLabelFormat: 'd MMM HH:mm',